requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
asyncssh==2.19.0
atpublic==5.1
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.2.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.22.7
dvc==3.59.0
dvc-data==3.16.8
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
flatten-dict==0.4.2
flufl-lock==8.1.0
fonttools==4.55.6
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
funcy==2.0
gitdb==4.0.12
gitpython==3.1.44
grandalf==0.8
gto==1.7.2
gudhi==3.10.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
iterative-telemetry==0.0.9
jedi==0.19.2
jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.3.4
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
kiwisolver==1.4.8
kombu==5.4.2
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook-shim==0.2.4
numpy==2.2.2
omegaconf==2.3.0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prometheus-client==0.21.1
prompt-toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydot==3.0.4
pygit2==1.17.0
pygments==2.19.1
pygtrie==2.5.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
referencing==0.36.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
ruamel-yaml==0.18.10
ruamel-yaml-clib==0.2.12
scipy==1.15.1
scmrepo==3.3.9
seaborn==0.13.2
semver==3.0.4
send2trash==1.8.3
setuptools==75.8.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqltrie==0.11.1
stack-data==0.6.3
tabulate==0.9.0
terminado==0.18.1
tinycss2==1.4.0
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
yarl==1.18.3
zc-lockfile==3.0.post1
